requirements.txt
# Web Framework Options
flask==3.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Testing
pytest==7.4.3
requests==2.31.0

# Development
python-dotenv==1.0.0

# Database (optional)
sqlalchemy==2.0.23

# Additional utilities
pydantic==2.5.0
